Attempt to fix Bug #7820 Does not backup a file greater than 2 GB Report from taps23@yahoo.com Changed legnth_of_file to off_t for ftello support and all relative changes made as well in various functions to be consistent

+++ branches/stable/mondo/mondo/common/libmondo-archive.c
@@ -1363,9 +1363,9 @@ hexdump | tr -s ' ' '0' | head -n1"));
 	wipe_archives(bkpinfo->scratchdir);
 	mvaddstr_and_log_it(g_currentY++, 74, "Done.");
 	if (IS_THIS_A_STREAMING_BACKUP(bkpinfo->backup_media_type)) {
-		write_header_block_to_stream(0, "start-of-tape",
+		write_header_block_to_stream((off_t)0, "start-of-tape",
 									 BLK_START_OF_TAPE);
-		write_header_block_to_stream(0, "start-of-backup",
+		write_header_block_to_stream((off_t)0, "start-of-backup",
 									 BLK_START_OF_BACKUP);
 	}
 	paranoid_free(command);
@@ -2174,8 +2174,7 @@ make_slices_and_images(struct s_bkpinfo *bkpinfo, char *biggielist_fname)
 	FILE *ftmp = NULL;
 	bool delete_when_done;
 	bool use_ntfsprog;
-	/*@ long long ****************************************** */
-	long long biggie_fsize;
+	off_t biggie_fsize;
 
 	assert(bkpinfo != NULL);
 	assert_string_is_neither_NULL_nor_zerolength(biggielist_fname);
@@ -2273,7 +2272,7 @@ make_slices_and_images(struct s_bkpinfo *bkpinfo, char *biggielist_fname)
 								  ntfsprog_fifo, biggie_file_number,
 								  noof_biggie_files, use_ntfsprog);
 			if (IS_THIS_A_STREAMING_BACKUP(bkpinfo->backup_media_type)) {
-				write_header_block_to_stream(0,
+				write_header_block_to_stream((off_t)0,
 											 calc_checksum_of_file
 											 (bigfile_fname),
 											 BLK_STOP_A_BIGGIE);
@@ -2460,7 +2459,7 @@ int make_those_afios_phase(struct s_bkpinfo *bkpinfo)
 						"Archiving regular files to media          ");
 
 	if (IS_THIS_A_STREAMING_BACKUP(bkpinfo->backup_media_type)) {
-		write_header_block_to_stream(0, "start-of-afioballs",
+		write_header_block_to_stream((off_t)0, "start-of-afioballs",
 									 BLK_START_AFIOBALLS);
 #if __FreeBSD__ == 5
 		log_msg(1,
@@ -2469,7 +2468,7 @@ int make_those_afios_phase(struct s_bkpinfo *bkpinfo)
 #else
 		res = make_afioballs_and_images_OLD(bkpinfo);
 #endif
-		write_header_block_to_stream(0, "stop-afioballs",
+		write_header_block_to_stream((off_t)0, "stop-afioballs",
 									 BLK_STOP_AFIOBALLS);
 	} else {
 		res = make_afioballs_and_images(bkpinfo);
@@ -2542,11 +2541,11 @@ int make_those_slices_phase(struct s_bkpinfo *bkpinfo)
 	if (IS_THIS_A_STREAMING_BACKUP(bkpinfo->backup_media_type)) {
 		res += write_EXAT_files_to_tape(bkpinfo, xattr_fname, acl_fname);
 		sprintf(blah, "%ld", count_lines_in_file(biggielist));
-		write_header_block_to_stream(0, blah, BLK_START_BIGGIEFILES);
+		write_header_block_to_stream((off_t)0, blah, BLK_START_BIGGIEFILES);
 	}
 	res = make_slices_and_images(bkpinfo, biggielist);
 	if (IS_THIS_A_STREAMING_BACKUP(bkpinfo->backup_media_type)) {
-		write_header_block_to_stream(0, "end-of-biggiefiles",
+		write_header_block_to_stream((off_t)0, "end-of-biggiefiles",
 									 BLK_STOP_BIGGIEFILES);
 	}
 	retval += res;
@@ -2930,7 +2929,7 @@ _move_files_to_stream(struct s_bkpinfo *bkpinfo, char *files_to_add, ...)
 	char stop_chr;
 	char *curr_file, *cf;
 	/*@ long long ****************************************************** */
-	long long length_of_incoming_file = 0;
+	off_t length_of_incoming_file = (off_t)0;
 	t_archtype type;
 	va_list ap;
 
@@ -2965,7 +2964,7 @@ _move_files_to_stream(struct s_bkpinfo *bkpinfo, char *files_to_add, ...)
 		retval += res;
 		unlink(curr_file);
 /* write closing header */
-		write_header_block_to_stream(0, "finished-writing-file", stop_chr);
+		write_header_block_to_stream((off_t)0, "finished-writing-file", stop_chr);
 	}
 	va_end(ap);
 
@@ -3305,9 +3304,9 @@ slice_up_file_etc(struct s_bkpinfo *bkpinfo, char *biggie_filename,
 	char *suffix;				// for compressed slices
 
 	/*@ long long ************************************************** */
-	long long totalread = 0;
-	long long totallength = 0;
-	long long length;
+	off_t totalread = (off_t)0;
+	off_t totallength = (off_t)0;
+	off_t length;
 
 	/*@ int ******************************************************** */
 	int retval = 0;
@@ -3359,7 +3358,7 @@ slice_up_file_etc(struct s_bkpinfo *bkpinfo, char *biggie_filename,
 		log_it("command = %s", command);
 		strcpy (tmp, call_program_and_get_last_line_of_output(command));
 		log_it("res of it = %s", tmp); 
-		totallength = atoll(tmp);
+		totallength = (off_t)atoll(tmp);
 	} else {
 		file_to_openin = biggie_filename;
 		sprintf(command, "md5sum '%s'", biggie_filename);
